requirements.txt
PyYAML>=6.0
ruamel.yaml>=0.18
requests>=2.31
